Create CODE_OF_CONDUCT.md
We want to be inclusive to the largest number of participants, with the most varied and diverse backgrounds possible. As such, we should take position as to how we will defend our contributors. Better settle on how we handle dire situations before they happen.

Nota: the code of conduct isn't clear on the disclosure of an incident per se. That should be investigated further.