change license to licence
Change "license" to "licence" for consistency in the user interface in English. 

License is both a noun and a verb in the United States.
If you live in any other English-speaking country, you will spell it "licence" when you use it as a noun and "license" when you use it as a verb.