Update iframe allowfullscreen to current method
The current PeerTube frontend generates iframe code that contains a old method
of allowing fullscreen: `allowfullscreen=""`. This method is surpassed by the
`allow="fullscreen"` property. This commit changes the rendered output to the
new code. I have tested it in Gitpod and the change worked as expected.

More documentation on this property:
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe#attr-allowfullscreen

Signed-off-by: Nico Rikken <nico@nicorikken.eu>