Ensure peertube root directory is setup to be traversed by nginx (#5028)
* Update production.md

On Ubuntu 22.04, the dir `/var/www/peertube` is restricted to user peertube & group peertube. This causes issues with many of the subsequent sudo commands, and also results in white-screen on initial visit to the site due to JS files not being accessible. This PR loosens those requirements, but maybe would be better if nginx is just added to peertube group? This is my first time installing PeertTube, so I'll defer to someone with more knowledge of this framework to vet the approach here.

* Set specific mode on peertube root directory

* Minor fixup

* Prefer ls instead of chmod to check directory

Co-authored-by: Chocobozzz <me@florianbigard.com>