Update privacy-concerns.component.html
When translated, the white space after "Reminder:" is cut off, so in other languages there is no white space between the "Reminder:" and the following text, which looks ugly (e.g. in the German translation). A solution is to move the space behind the translated phrase (using the html-entity for space &#32).