Add the ability to specify peertube UID and GID via environment variables
This adds support for specifying the peertube user and group IDs using the standard environment variable pattern similar to linuxserver.io images.  This allows for better external permission management, specifically this makes it easier to support mapping volumes to NFS shares using nfs4 id mapping.

This changes elects *not* to use the linuxserver.io variable names of PUID and PGID since the peertube container ends up having both a "node" and "peertube" user created and it would not be obvious which of those PUID/PGID should apply to.