API-access for external auth plugins
This allows auth plugins to return the bypass token as a JSON
response, rather than in a redirection.

Using a redirect and putting the bypass token in the query
parameter works when the browser goes through the login
flow using the PeerTube frontend and redirect back to an
external frontend afterwards. But for a purely API based
access from an external frontend, the redirect does not work.

See https://framagit.org/framasoft/peertube/official-plugins/-/merge_requests/33
for more details on why it does not work and how the
proposed change can be used in a plugin.